+++ app/src/main/java/br/com/rbrthmn/ui/financialcompanion/navigation/ComFinNavGraph.kt
@@ -27,8 +27,8 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import br.com.rbrthmn.ui.financialcompanion.screens.home.HomeDestination
 import br.com.rbrthmn.ui.financialcompanion.screens.home.HomeScreen
-import br.com.rbrthmn.ui.financialcompanion.screens.IncomeDivisionsDestination
-import br.com.rbrthmn.ui.financialcompanion.screens.IncomeDivisionsScreen
+import br.com.rbrthmn.ui.financialcompanion.screens.incomedivisions.IncomeDivisionsDestination
+import br.com.rbrthmn.ui.financialcompanion.screens.incomedivisions.IncomeDivisionsScreen
 import br.com.rbrthmn.ui.financialcompanion.screens.MoreFeaturesDestination
 import br.com.rbrthmn.ui.financialcompanion.screens.MoreFeaturesScreen
 import br.com.rbrthmn.ui.financialcompanion.screens.OperationsDestination
